+++ src/vis/index.ts
@@ -1,3 +1,3 @@
 export * from './provvis';
 export * from './storyvis';
-export * from './vis_loader';
+export * from './VisLoader';
