+++ src/wrapper/index.ts
@@ -1,3 +1,3 @@
 export * from './ACLUEWrapper';
 export * from './CLUEWrapper';
-export * from './wrapperUtils';
+export * from './WrapperUtils';
